fix: image via url not working
